[S09P22B310-198] fix: 멤버 엔티티 수정

See merge request s09-fintech-finance-sub2/S09P22B310!115

+++ backend/src/main/java/com/ntt/mwonimoney/domain/member/entity/Guest.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class Guest extends Member {
 
 	@Builder
-	public Guest(byte status, String name, String nickname, String birthday, SocialProvider socialProvider,
+	public Guest(int status, String name, String nickname, String birthday, SocialProvider socialProvider,
 		String socialId, String email) {
 		super(status, name, nickname, birthday, socialProvider, socialId, email, MemberRole.GUEST);
 	}
